+++ .env
@@ -1,0 +1,6 @@
+DATABASE_URL="mongodb+srv://bambicom:92oUMAN20RqVPodF@bambi.qbpgzdo.mongodb.net/development"
+
+NEXTAUTH_JWT_SECRET=xhQXDfA1YEvnwBig1gMXNlwucwzLb8vG3RalPx7m9wg=
+
+NEXTAUTH_SECRET=wpWcpZBQs0nOON6wL9V3YSjLmsMgeQWXXfjZSThj7Lg=
+AUTH_SECRET=wpWcpZBQs0nOON6wL9V3YSjLmsMgeQWXXfjZSThj7Lg=
