+++ src/lib/navData.ts
@@ -1,6 +1,6 @@
 export const pagesArray = [
   { title: "About Us", href: "/about" },
   { title: "Contact Us", href: "/contact" },
-  { title: "Login", href: "/auth/login" },
-  { title: "Register", href: "/auth/register" },
+  // { title: "Login", href: "/auth/login" },
+  // { title: "Register", href: "/auth/register" },
 ];
