+++ src/types.ts
@@ -16,7 +16,6 @@ export type ProductsType = Products[];
 
 export type CartType = {
   id: string | number;
-  uid?: string;
   src: string;
   href: string;
   title: string;
